fix：第6批低修改风险findings
